+++ CHANGELOG.md
@@ -71,6 +71,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
   chore(rollup): add missing fns in controller
 * [#569](https://github.com/babylonlabs-io/finality-provider/pull/569)
   chore(rollup): add highest voted height query
+* [#567](https://github.com/babylonlabs-io/finality-provider/pull/567) chore:
+  e2e cosmos bsn
+* [#570](https://github.com/babylonlabs-io/finality-provider/pull/570)
+  chore(rollup): don't allow FP program to start if not in allowlist
 * [#572](https://github.com/babylonlabs-io/finality-provider/pull/572)
   chore(rollup): ensure pub randomness is timestamped